+++ src/test/converter2/issues/gh2683.ts
@@ -1,0 +1,17 @@
+export declare function action(
+    /**
+     * Param comment
+     * @param data - Data
+     */
+    callback: (data: {}) => void,
+): void;
+
+/**
+ * @param callback Param comment2
+ */
+export declare function action2(
+    /**
+     * @param data - Data2
+     */
+    callback: (data: {}) => void,
+): void;