+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 -   Fixed an issue where links in packages mode would be resolved incorrectly, #2680.
+-   Fixed handling of `@param` tags on comments attached to function callback parameters, #2683.
 
 ## v0.26.6 (2024-08-18)